+++ NEWS.md
@@ -1,5 +1,7 @@
 # reprex (development version)
 
+* New `local_figures` argument for `reprex()` and `reprex_document()` keeps figures as local files instead of uploading to imgur (#673, @BjarkeHautop).
+
 # reprex 2.1.1
 
 * `reprex(style = FALSE)` will never nag about installing styler (#461).